Move `Allocation` to separate module
