+++ rust/relay/src/lib.rs
@@ -1,3 +1,4 @@
+mod allocation;
 mod auth;
 mod rfc8656;
 mod server;
@@ -9,6 +10,7 @@ mod udp_socket;
 #[cfg(feature = "proptest")]
 pub mod proptest;
 
+pub use allocation::Allocation;
 pub use server::{
     Allocate, AllocationId, Attribute, Binding, ChannelBind, ChannelData, ClientMessage, Command,
     CreatePermission, Refresh, Server,
